refactor(action): split action module into separate files
