+++ src/action/effect.rs
@@ -1,11 +1,4 @@
-use thiserror::Error;
-
-#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Copy)]
-pub enum Position {
-    Desperate,
-    Risky,
-    Controlled,
-}
+use super::{ActionError, Position, Result};
 
 #[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Copy)]
 pub enum Effect {
@@ -15,20 +8,6 @@ pub enum Effect {
     Great,
     Extreme,
 }
-
-#[derive(Error, Debug, PartialEq)]
-pub enum ActionError {
-    #[error("cannot decrease position below {0:?}")]
-    PositionClampedLow(Position),
-    #[error("cannot increase position above {0:?}")]
-    PositionClampedHigh(Position),
-    #[error("cannot increase effect above {0:?}")]
-    EffectClampedHigh(Effect),
-    #[error("cannot decrease effect below {0:?}")]
-    EffectClampedLow(Effect),
-}
-
-type Result<T> = std::result::Result<T, ActionError>;
 
 impl Effect {
     pub fn increase(&self) -> Self {
@@ -72,58 +51,12 @@ impl Effect {
     }
 }
 
-impl Position {
-    pub fn improve(&self) -> Self {
-        match self {
-            Position::Desperate => Position::Risky,
-            Position::Risky => Position::Controlled,
-            Position::Controlled => Position::Controlled,
-        }
-    }
-
-    pub fn diminish(&self) -> Self {
-        match self {
-            Position::Controlled => Position::Risky,
-            Position::Risky => Position::Desperate,
-            Position::Desperate => Position::Desperate,
-        }
-    }
-
-    pub fn trade_for_effect(&self, effect: Effect) -> Result<(Self, Effect)> {
-        if *self == Position::Desperate {
-            return Err(ActionError::PositionClampedLow(Position::Desperate));
-        }
-
-        if effect >= Effect::Great {
-            return Err(ActionError::EffectClampedHigh(Effect::Great));
-        }
-
-        Ok((self.diminish(), effect.increase()))
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use proptest::prelude::*;
     use rstest::rstest;
 
     use super::*;
-
-    #[rstest]
-    #[case::from_desperate_to_risky(Position::Desperate, Position::Risky)]
-    #[case::from_risky_to_controlled(Position::Risky, Position::Controlled)]
-    #[case::from_controlled_to_controlled(Position::Controlled, Position::Controlled)]
-    fn test_position_improves(#[case] from: Position, #[case] to: Position) {
-        assert_eq!(from.improve(), to);
-    }
-
-    #[rstest]
-    #[case::from_controlled_to_risky(Position::Controlled, Position::Risky)]
-    #[case::from_risky_to_desperate(Position::Risky, Position::Desperate)]
-    #[case::from_desperate_to_desperate(Position::Desperate, Position::Desperate)]
-    fn test_position_diminishes(#[case] from: Position, #[case] to: Position) {
-        assert_eq!(from.diminish(), to)
-    }
 
     #[rstest]
     #[case::from_zero_to_limited(Effect::Zero, Effect::Limited)]
@@ -163,29 +96,6 @@ mod tests {
     #[case::from_extreme_to_extreme(Effect::Extreme, Effect::Extreme, Effect::Extreme)]
     fn test_at_most_edge_cases(#[case] from: Effect, #[case] clamp: Effect, #[case] to: Effect) {
         assert_eq!(from.at_most(clamp), to);
-    }
-
-    #[rstest]
-    #[case::controlled_to_risky_for_effect(Position::Controlled, Effect::Limited, Position::Risky, Effect::Standard)]
-    #[case::risky_to_desperate_for_effect(Position::Risky, Effect::Limited, Position::Desperate, Effect::Standard)]
-    #[case::controlled_to_risky_for_great_effect(Position::Controlled, Effect::Standard, Position::Risky, Effect::Great)]
-    fn test_trade_position_for_effect(
-        #[case] initial_position: Position, #[case] initial_effect: Effect, #[case] expected_position: Position, #[case] expected_effect: Effect,
-    ) {
-        let (new_position, new_effect) = initial_position
-            .trade_for_effect(initial_effect)
-            .expect("should have traded successfully");
-        assert_eq!(new_position, expected_position);
-        assert_eq!(new_effect, expected_effect);
-    }
-
-    #[rstest]
-    #[case::cannot_decrease_below_desperate(Position::Desperate, Effect::Limited, ActionError::PositionClampedLow(Position::Desperate))]
-    #[case::cannot_increase_above_great(Position::Controlled, Effect::Great, ActionError::EffectClampedHigh(Effect::Great))]
-    #[case::cannot_increase_above_extreme(Position::Controlled, Effect::Extreme, ActionError::EffectClampedHigh(Effect::Great))]
-    fn test_fail_to_trade_position_for_effect(#[case] initial_position: Position, #[case] initial_effect: Effect, #[case] error: ActionError) {
-        let err = initial_position.trade_for_effect(initial_effect).expect_err("should have failed");
-        assert_eq!(err, error);
     }
 
     #[rstest]
